Refactor: split four large files into focused modules
- routes/trainable_models.py (828 → 426 lines): extract /api/models/registry/*
  into routes/models_registry.py. Trainable model store and in-memory model
  registry are distinct resources sharing a blueprint.
- routes/datasets.py (952 → 693 lines): extract /api/datasets/registry/*
  into routes/datasets_registry.py.
- datasets/loader.py (1244 → 196 lines): split into loader_folder.py
  (folder loaders), loader_pickle.py (pickle loaders + sidecars + image
  embed), loader_demo.py (demo loader). loader.py becomes the public façade,
  re-exporting all symbols and holding shared private helpers + export.
- settings.py (855 → 716 lines): extract factory functions
  (make_accessors, clamp, one_of, make_per_side_setting) into
  settings_factory.py. Generation loops stay in settings.py since they
  install accessors into the module's namespace.

All in-repo callers and test patches updated.
