+++ app.py
@@ -49,10 +49,12 @@ from vtsearch.routes import (  # noqa: E402
     media_server_bp,
     medias_bp,
     datasets_bp,
+    datasets_registry_bp,
     detectors_bp,
     exporters_bp,
     label_importers_bp,
     main_bp,
+    models_registry_bp,
     processor_importers_bp,
     settings_bp,
     settings_io_bp,
@@ -172,6 +174,7 @@ app.register_blueprint(medias_bp)
 app.register_blueprint(sorting_bp)
 app.register_blueprint(detectors_bp)
 app.register_blueprint(datasets_bp)
+app.register_blueprint(datasets_registry_bp)
 app.register_blueprint(exporters_bp)
 app.register_blueprint(label_importers_bp)
 app.register_blueprint(processor_importers_bp)
@@ -179,6 +182,7 @@ app.register_blueprint(settings_bp)
 app.register_blueprint(settings_io_bp)
 app.register_blueprint(sync_sources_bp)
 app.register_blueprint(trainable_models_bp)
+app.register_blueprint(models_registry_bp)
 
 
 # ---------------------------------------------------------------------------
