+++ vtsearch/routes/__init__.py
@@ -7,10 +7,12 @@ from vtsearch.routes.labels import labels_bp
 from vtsearch.routes.media_server import media_server_bp
 from vtsearch.routes.medias import medias_bp
 from vtsearch.routes.datasets import datasets_bp
+from vtsearch.routes.datasets_registry import datasets_registry_bp
 from vtsearch.routes.detectors import detectors_bp
 from vtsearch.routes.exporters import exporters_bp
 from vtsearch.routes.label_importers import label_importers_bp
 from vtsearch.routes.main import main_bp
+from vtsearch.routes.models_registry import models_registry_bp
 from vtsearch.routes.processor_importers import processor_importers_bp
 from vtsearch.routes.settings import settings_bp
 from vtsearch.routes.settings_io import settings_io_bp
@@ -30,8 +32,10 @@ __all__ = [
     "sync_sources_bp",
     "detectors_bp",
     "datasets_bp",
+    "datasets_registry_bp",
     "exporters_bp",
     "label_importers_bp",
+    "models_registry_bp",
     "processor_importers_bp",
     "settings_bp",
     "settings_io_bp",
